+++ tests/testthat/test-plot.R
@@ -41,7 +41,7 @@ test_that("Pedigree plotting test", {
     )
 
     lst <- ped_to_plotdf(pedi, precision = 4)
-    expect_equal(length(lst), 2)
+    expect_equal(length(lst), 3)
     expect_equal(dim(lst$df), c(82, 17))
     expect_snapshot(lst)
     expect_equal(
@@ -62,8 +62,14 @@ test_that("Pedigree fails to line up", {
     df1 <- sampleped[sampleped$famid == "1", ]
     ped1 <- Pedigree(df1)
     vdiffr::expect_doppelganger("ped1",
-        function() plot(ped1, precision = 4)
+        function() suppressWarnings(plot(ped1, precision = 4))
     )
+
+    expect_equal(
+        suppressWarnings(plot(ped1, precision = 4))$ind_not_plot,
+        c("1_113")
+    )
+
     # With reordering it's better
     df1reord <- df1[c(35:41, 1:34), ]
     ped1reord <- Pedigree(df1reord)